feat: add hotel management functionality

+++ be/holidate/src/main/java/com/webapp/holidate/dto/response/location/entertainment_venue/EntertainmentVenueResponse.java
@@ -16,5 +16,5 @@ import lombok.experimental.FieldDefaults;
 public class EntertainmentVenueResponse {
   String id;
   String name;
-  int distance;
+  double distance;
 }
